Update home_screen.dart
I added a new background image which can be found in the UI images branch. I also changed some of the coloring. I will text in the discord for you guys to run it and see what you think! 

I also had to edit the pubspec.yaml just inluding the assets image